scheduling/server/pyproject.toml
[tool.poetry]
name = "scheduling"
version = "0.1.0"
description = ""
authors = ["Simon <berndts.tin21@student.dhbw-heidenheim.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
schedule = "^1.2.0"
redis = "^4.6.0"
cassandra-driver = "^3.28.0"
pydantic = "^2.3.0"
python-socketio = "^5.9.0"
requests = "^2.31.0"
websocket-client = "^1.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
